// frontend/lib/api/generated/model/almacenOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface AlmacenOut {
  almacen_id: number;
  codigo: string;
  nombre: string;
  sede_id: number;
  direccion: string | null;
  tipo_comedor: string;
  responsable_id: number;
  estado: string;
  creado_en: string;
}